First commit
Committing to save

This isn't anywhere near functional, but it's a darn sight better
than nothing.

Choosing to commit here because this is the point where I realized
that I really need to encode bought_so_far in gamestate, because
it's so important for RandomStrategy. And I don't want to lose
what I've done so far.