First "working" prototype
This code allows the two randomStrategies to play
a game against each other, and records the results.

However, shortly after completing this, I realized that
this model wasn't really extensible. Situations where you
have to make a special kind of choice - when playing a Thief,
a Spy, or a Library, for instance - will require specialized
kinds of representation of the gamestate, and have different
possible actions/responses.

The solution, I believe, is for a Gamestate representation
to be specific to the situation. There would be, for instance,
a different input/output vector formulation for:
* It is the Buy Phase - what do you do?
* It is the Action Phase - what do you do?
* You are playing a Harbinger - do you return a card?
* You are playing a Chancellor - do you put your deck into your discard?
* Your opponent played a Swindler - what card do you gain?
* Your opponent played an Attack card - do you reveal a Reaction card?

This would suggest that the Strategy should then have multiple
networks - one for each possible Situation. Some (like the
"base Buy/Action Phase" ones) will naturally be exercised more
often than others!

It will be important to train each Persona (collection of networks)
only on data generated by itself (that is, not to amalgamate data
of "all actions taken by a Situational-network" and to train that
Situational-network on the combination of them) - the value of
actions is very dependent on the likelihood of other actions (e.g.
biasing in favour of buying one type of card is more valuable if you
are also biased in favour of playing it, as a very simple example)